chore(nuz-cli): use common utils in build and dev commands
